requirements.txt
# Core dependencies
anthropic>=0.39.0
azure-ai-formrecognizer>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# CLI and UI
click>=8.1.0
rich>=13.0.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0  # Para Excel export
python-dateutil>=2.8.0

# Utils
python-json-logger>=2.0.0
requests>=2.31.0
tenacity>=8.2.0  # Para retry logic

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.1.0
